Moved: distillation and activation_sparsity modifiers
